+++ frontend/src/javascript/codeParser/CodeParser.js
@@ -1,3 +1,5 @@
+import { CodeParserUtils } from "./CodeParserUtils";
+
 export class CodeParser {
     actions = {
         "<algodebug-variable>": function (parent, tag) {
@@ -92,103 +94,16 @@ export class CodeParser {
     }
 
     prepareCode() {
-        this.code = CodeUtils.insertVariableTags(this.code, this.variables);
-        this.code = CodeUtils.insertBreakpointTags(this.code, this.breakpoints);
+        this.code = CodeParserUtils.insertVariableTags(this.code, this.variables);
+        this.code = CodeParserUtils.insertBreakpointTags(this.code, this.breakpoints);
     }
 
     parseCode() {
-        this.code = CodeUtils.removeVariableTags(this.code);
-        this.code = CodeUtils.replaceBreakpointTags(this.code, this.parsedBreakpoints);
-        this.code = CodeUtils.insertConvertersAfterIncludes(this.code, this.converters);
-        this.code = CodeUtils.insertConvertersAtTheEnd(this.code, this.converters);
-        this.code = CodeUtils.insertAlgodebugMacros(this.code);
-        this.code = CodeUtils.insertNecessaryIncludes(this.code);
-    }
-}
-
-class CodeUtils {
-    static insertVariableTags(code, variables) {
-        for (let variable of variables.sortedBy("start", -1)) {
-            code =
-                code.slice(0, variable.start) +
-                "<algodebug-variable>" +
-                variable.id +
-                "</algodebug-variable>" +
-                code.slice(variable.end);
-        }
-        return code;
-    }
-
-    static insertBreakpointTags(code, breakpoints) {
-        let lines = code.split("\n");
-        for (let breakpoint of breakpoints.sortedBy("id", -1)) {
-            lines[breakpoint.id] += `<algodebug-breakpoint>${breakpoint.id}</algodebug-breakpoint>`;
-        }
-        code = lines.join("\n");
-        return code;
-    }
-
-    static removeVariableTags(code) {
-        let i = -1;
-        while ((i = code.indexOf("<algodebug-variable>")) > 0) {
-            let start = i + "<algodebug-variable>".length;
-            let end = code.indexOf("</algodebug-variable>", start);
-            let variableID = code.substring(start, end);
-            let variableName = variableID.split("@")[0];
-            code = code.substring(0, i) + variableName + code.substring(end + "</algodebug-variable>".length);
-        }
-        return code;
-    }
-
-    static replaceBreakpointTags(code, parsedBreakpoints) {
-        for (let breakpoint of parsedBreakpoints) {
-            let variables = breakpoint.variables
-                .map((variable) => `ALGODEBUG_VARIABLE(${variable.name}, ${variable.id.split("@")[1]})`)
-                .join(" << ");
-            code = code.replace(
-                `<algodebug-breakpoint>${breakpoint.line}</algodebug-breakpoint>`,
-                variables.length !== 0
-                    ? ` ALGODEBUG_BREAKPOINT(${breakpoint.line}, ${variables});`
-                    : ` ALGODEBUG_EMPTY_BREAKPOINT(${breakpoint.line});`
-            );
-        }
-        return code;
-    }
-
-    static insertAlgodebugMacros(code) {
-        return (
-            `#define ALGODEBUG_VARIABLE(x, y) "<algodebug-variable " << "name=\\"" << #x << "@" << y << "\\">\\n" << x << "\\n</algodebug-variable>\\n"\n` +
-            `#define ALGODEBUG_BREAKPOINT(line, x) std::cout << "<algodebug-breakpoint " << "line=\\"" << line << "\\">\\n" << x << "</algodebug-breakpoint>\\n"\n` +
-            `#define ALGODEBUG_EMPTY_BREAKPOINT(line) std::cout << "<algodebug-breakpoint " << "line=\\"" << line << "\\">\\n</algodebug-breakpoint>\\n"\n\n` +
-            code
-        );
-    }
-
-    static insertConvertersAfterIncludes(code, converters) {
-        converters = converters
-            .map((converter) => converter.code.slice(0, converter.code.indexOf("{")).trim() + ";")
-            .join("\n");
-
-        let includeStartPosition = code.lastIndexOf("#include");
-        let includeEndPosition = code.indexOf(">", includeStartPosition);
-        let usingNamespaceStartPosition = code.lastIndexOf("using namespace");
-        let usingNamespaceEndPosition = code.indexOf(";", usingNamespaceStartPosition);
-
-        let position = Math.max(includeEndPosition, usingNamespaceEndPosition) + 1;
-        code = code.slice(0, position) + "\n\n" + converters + code.slice(position);
-        return code;
-    }
-
-    static insertConvertersAtTheEnd(code, converters) {
-        converters = converters.map((converter) => converter.code).join("\n\n");
-        return code + "\n\n" + converters;
-    }
-
-    static insertNecessaryIncludes(code) {
-        const regex = /#include[ \t]<iostream>/g;
-        if (!regex.test(code)) {
-            return "#include <iostream>\n" + code;
-        }
-        return code;
+        this.code = CodeParserUtils.removeVariableTags(this.code);
+        this.code = CodeParserUtils.replaceBreakpointTags(this.code, this.parsedBreakpoints);
+        this.code = CodeParserUtils.insertConvertersAfterIncludes(this.code, this.converters);
+        this.code = CodeParserUtils.insertConvertersAtTheEnd(this.code, this.converters);
+        this.code = CodeParserUtils.insertAlgodebugMacros(this.code);
+        this.code = CodeParserUtils.insertNecessaryIncludes(this.code);
     }
 }
